edit-product-gallery

+++ app/Http/Livewire/Admin/AdminProductComponent.php
@@ -22,7 +22,9 @@ class AdminProductComponent extends Component
             $images = explode(",",$product->images);
             foreach ($images as $image)
             {
-                unlink('assets/images/products'.'/'.$image);
+                if($images){
+                    unlink('assets/images/products'.'/'.$image);
+                }
             }
         }
         $product->delete();
